@@ -4,43 +4,40 @@ description: Library for issuing commands via blossm
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2"
+  sdk: ^3.7.2
+
+publish_to: none
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.0+3
-  device_info: ^0.4.0+1
+  http: ^1.4.0
+  device_info: ^2.0.3
   blossm_command:
-    path: ../../blossm_command
+    git:
+      url: git@github.com:mejango/experiment.git
+      ref: main
+      path: components/blossm_command
   secure_storage:
     git:
-      url: git@github.com:RoofVillage/roof-x.git
-      ref: master
-      path: lib/secure_storage
+      url: git@github.com:mejango/experiment.git
+      ref: main
+      path: components/lib/secure_storage
   navigator:
     git:
-      url: git@github.com:RoofVillage/roof-x.git
-      ref: master
-      path: lib/navigator
-  welcome_artboard:
-    git:
-      url: git@github.com:RoofVillage/roof-x.git
-      ref: master
-      path: sustainers_market/artboards/welcome
-  auth_verify_code_artboard:
-    git:
-      url: git@github.com:RoofVillage/roof-x.git
-      ref: master
-      path: sustainers_market/artboards/auth_verify_code
-
-dependency_overrides:
-  welcome_artboard:
-    path: ../artboards/welcome
-  register_artboard:
-    path: ../artboards/register
-  auth_verify_code_artboard:
-    path: ../artboards/auth_verify_code
+      url: git@github.com:mejango/experiment.git
+      ref: main
+      path: components/lib/navigator
+  # welcome_artboard:
+  #   git:
+  #     url: git@github.com:mejango/experiment.git
+  #     ref: main
+  #     path: apps/app_1/artboards/welcome
+  # auth_verify_code_artboard:
+  #   git:
+  #     url: git@github.com:mejango/experiment.git
+  #     ref: main
+  #     path: apps/app_1/artboards/auth_verify_code
 
 dev_dependencies:
   flutter_test:
