@@ -1,8 +1,8 @@
-import 'package:auth_verify_code_artboard/index.dart';
+// import 'package:auth_verify_code_artboard/index.dart';
 import 'package:blossm_command/index.dart';
-import 'package:flutter/material.dart';
-import 'package:navigator/index.dart';
-import 'package:welcome_artboard/index.dart';
+// import 'package:flutter/material.dart';
+// import 'package:navigator/index.dart';
+// import 'package:welcome_artboard/index.dart';
 
 import 'token_storer.dart';
 
@@ -11,13 +11,13 @@ mixin DispatcherConfig {
 
   final TokenStore tokenStore = SMTokenStore();
 
-  final Function(BuildContext) redirectToWelcome =
-      (BuildContext context) => ArtboardNavigator.of(context).goTo(
-            WelcomeVerticalFullScreenArtboard(),
-          );
+  // final Function(BuildContext) redirectToWelcome =
+  //     (BuildContext context) => ArtboardNavigator.of(context).goTo(
+  //           WelcomeVerticalFullScreenArtboard(),
+  //         );
 
-  final Function(BuildContext) redirectToChallengeAnswer =
-      (BuildContext context) => ArtboardNavigator.of(context).goTo(
-            AuthVerifyCodeVerticalFloatingArtboard(),
-          );
+  // final Function(BuildContext) redirectToChallengeAnswer =
+  //     (BuildContext context) => ArtboardNavigator.of(context).goTo(
+  //           AuthVerifyCodeVerticalFloatingArtboard(),
+  //         );
 }
